requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.4
anthropic==0.42.0
chromadb==0.5.23
python-dotenv==1.0.1
